# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: LiveQs V1 API
  version: 0.1.0
  description: >-
    Authoritative HTTP contract for the single-Owner LiveQs V1 service. Event
    bodies are governed by the versioned JSON Schemas under schemas/.
servers:
  - url: http://localhost:3000
    description: Local and trusted-LAN development
tags:
  - name: Owner
    description: Single-Owner authentication and revocable browser session lifecycle.
  - name: Credentials
    description: >-
      Owner-managed Device Tokens and Query Tokens. Each credential carries an
      actor type, scopes, permitted event types, a privacy ceiling, an optional
      expiry, and independently revocable state.
  - name: Events
    description: Versioned historical observations. Heartbeats are not events.
paths:
  /api/v1/owner/status:
    get:
      operationId: getOwnerStatus
      summary: Report whether the Owner password is configured
      description: >-
        Public endpoint used by the WebUI to distinguish first-run setup from
        login. There is exactly one implicit Owner and no username.
      tags: [Owner]
      security: []
      responses:
        "200":
          description: Initialization state of the single implicit Owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerStatus"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/owner/setup:
    post:
      operationId: setupOwnerPassword
      summary: Create the Owner password on an uninitialized instance
      description: >-
        Allowed only while no Owner credential exists. Establishes the first
        Owner session and sets the liveqs_session cookie on success.
      tags: [Owner]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerPasswordRequest"
      responses:
        "204":
          description: >-
            Owner credential created and first session established. The
            response sets the HttpOnly, SameSite liveqs_session cookie.
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/owner/login:
    post:
      operationId: loginOwner
      summary: Log in with the Owner password
      description: >-
        Verifies the password and establishes a revocable server-managed
        session in the HttpOnly, SameSite liveqs_session cookie.
      tags: [Owner]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OwnerPasswordRequest"
      responses:
        "204":
          description: Session established; the response sets the liveqs_session cookie.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/owner/logout:
    post:
      operationId: logoutOwner
      summary: Revoke the current Owner session
      description: >-
        Revokes the presented session server-side and clears the
        liveqs_session cookie. The revoked session can no longer access
        protected endpoints.
      tags: [Owner]
      security:
        - OwnerSession: []
      responses:
        "204":
          description: Session revoked and cookie cleared.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/owner/session:
    get:
      operationId: getOwnerSession
      summary: Probe the current Owner session
      description: WebUI entry point for distinguishing an authenticated browser from a guest.
      tags: [Owner]
      security:
        - OwnerSession: []
      responses:
        "200":
          description: The presented session is valid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerSessionInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/credentials:
    get:
      operationId: listCredentials
      summary: List Device Tokens and Query Tokens with lifecycle state
      description: >-
        Returns every credential of the single Owner with creation time,
        expiry, last use, and revocation state. Token material is never
        included; each credential is identified by its recognizable prefix.
      tags: [Credentials]
      security:
        - OwnerSession: []
      responses:
        "200":
          description: All credentials, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      operationId: createCredential
      summary: Create a Device Token or Query Token
      description: >-
        Creates a scoped credential for a collector or read-only agent. The
        plaintext token is returned exactly once in this response; the server
        persists only a secure digest and a recognizable prefix.
      tags: [Credentials]
      security:
        - OwnerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialCreateRequest"
      responses:
        "201":
          description: Credential created; token plaintext is shown only in this response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialCreated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/credentials/{credential_id}/revoke:
    post:
      operationId: revokeCredential
      summary: Revoke a credential immediately
      description: >-
        Marks the credential revoked so it is rejected from the next use
        onwards. Revoking an already revoked credential is idempotent.
      tags: [Credentials]
      security:
        - OwnerSession: []
      parameters:
        - $ref: "#/components/parameters/CredentialId"
      responses:
        "204":
          description: The credential is revoked and can no longer authenticate.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/events/batch:
    post:
      operationId: batchUpsertEvents
      summary: Submit event revisions with per-item acknowledgements
      tags: [Events]
      security:
        - DeviceToken: []
      x-required-scopes: [events:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventBatchRequest"
      responses:
        "200":
          description: Every submitted item has an independent acknowledgement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventBatchResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/events:
    get:
      operationId: listEvents
      summary: List latest valid event revisions
      tags: [Events]
      security:
        - OwnerSession: []
        - QueryToken: []
      x-required-scopes: [events:read]
      parameters:
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Timezone"
        - $ref: "#/components/parameters/EventType"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of latest, non-invalidated event revisions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
components:
  securitySchemes:
    OwnerSession:
      type: apiKey
      in: cookie
      name: liveqs_session
      description: Revocable HttpOnly Owner browser session.
    DeviceToken:
      type: http
      scheme: bearer
      bearerFormat: LiveQs device token
      description: >-
        Independently revocable upload-only collector credential. Plaintext
        tokens start with the recognizable prefix lqdev_ and are shown once at
        creation; the server stores only a SHA-256 digest.
    QueryToken:
      type: http
      scheme: bearer
      bearerFormat: LiveQs query token
      description: >-
        Scoped, expiring, read-only credential for AI and integrations.
        Plaintext tokens start with the recognizable prefix lqqry_ and are
        shown once at creation; the server stores only a SHA-256 digest.
  parameters:
    CredentialId:
      name: credential_id
      in: path
      required: true
      description: Public identifier of the credential, as returned at creation.
      schema:
        type: string
        minLength: 1
    From:
      name: from
      in: query
      required: true
      description: Inclusive UTC instant.
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      required: true
      description: Exclusive UTC instant; must be later than from.
      schema:
        type: string
        format: date-time
    Timezone:
      name: timezone
      in: query
      required: true
      description: IANA timezone used to interpret reporting boundaries.
      schema:
        type: string
        minLength: 1
        example: Asia/Shanghai
    EventType:
      name: event_type
      in: query
      required: false
      schema:
        type: string
        example: activity.interval
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque continuation cursor returned by the preceding page.
      schema:
        type: string
        minLength: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
  responses:
    BadRequest:
      description: Request shape or semantics are invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Credential is missing, expired, revoked, or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Credential lacks the required scope, event type, or privacy ceiling.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Credential rate limit was exceeded.
      headers:
        Retry-After:
          description: Seconds until another request may be attempted.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The request conflicts with existing state.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The referenced resource does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    OwnerStatus:
      title: Owner Status
      type: object
      additionalProperties: false
      required: [initialized]
      properties:
        initialized:
          type: boolean
          description: True once the single implicit Owner password has been created.
    OwnerPasswordRequest:
      title: Owner Password Request
      type: object
      additionalProperties: false
      required: [password]
      properties:
        password:
          type: string
          minLength: 8
          maxLength: 256
          description: The Owner password. Never logged or stored in plaintext.
    OwnerSessionInfo:
      title: Owner Session Info
      type: object
      additionalProperties: false
      required: [authenticated]
      properties:
        authenticated:
          type: boolean
          description: Always true when the endpoint responds 200.
    CredentialKind:
      title: Credential Kind
      type: string
      enum: [device_token, query_token]
      description: >-
        Actor type the credential authenticates: a device collector uploads
        events, a query agent reads approved data domains.
    CredentialScope:
      title: Credential Scope
      type: string
      enum: [events:write, events:read]
      description: >-
        Capability granted to the credential. Device tokens carry events:write
        and query tokens carry events:read; a credential never holds a scope
        outside its actor type.
    CredentialPrivacyCeiling:
      title: Credential Privacy Ceiling
      type: string
      enum: [normal, sensitive, private]
      description: >-
        Maximum privacy level the credential may touch, ordered normal <
        sensitive < private. Uploads above the ceiling are rejected per item;
        reads exclude data above the ceiling.
    CredentialCreateRequest:
      title: Credential Create Request
      type: object
      additionalProperties: false
      required: [kind, name, scopes]
      properties:
        kind:
          $ref: "#/components/schemas/CredentialKind"
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: Owner-chosen label, such as a collector or agent name.
        scopes:
          type: array
          minItems: 1
          maxItems: 8
          uniqueItems: true
          items:
            $ref: "#/components/schemas/CredentialScope"
        allowed_event_types:
          type: array
          default: []
          uniqueItems: true
          items:
            type: string
            pattern: "^[a-z][a-z0-9]*(\\.[a-z][a-z0-9]*)+$"
          description: >-
            Event types the credential may upload or read. Empty permits every
            registered event type; a non-empty list is an exact allow-list.
        privacy_ceiling:
          $ref: "#/components/schemas/CredentialPrivacyCeiling"
          default: normal
        expires_at:
          type: [string, "null"]
          format: date-time
          default: null
          description: >-
            UTC instant after which the credential is rejected at use; null
            means the credential never expires.
    CredentialView:
      title: Credential View
      type: object
      additionalProperties: false
      required:
        - credential_id
        - kind
        - name
        - token_prefix
        - scopes
        - allowed_event_types
        - privacy_ceiling
        - created_at
        - expires_at
        - last_used_at
        - revoked_at
      properties:
        credential_id:
          type: string
          minLength: 1
          description: Public identifier used to revoke the credential.
        kind:
          $ref: "#/components/schemas/CredentialKind"
        name:
          type: string
        token_prefix:
          type: string
          minLength: 1
          description: >-
            Recognizable prefix of the token plaintext for identification;
            never the full token or its digest.
        scopes:
          type: array
          items:
            $ref: "#/components/schemas/CredentialScope"
        allowed_event_types:
          type: array
          items:
            type: string
        privacy_ceiling:
          $ref: "#/components/schemas/CredentialPrivacyCeiling"
        created_at:
          type: string
          format: date-time
        expires_at:
          type: [string, "null"]
          format: date-time
        last_used_at:
          type: [string, "null"]
          format: date-time
          description: UTC instant of the most recent accepted use; null while unused.
        revoked_at:
          type: [string, "null"]
          format: date-time
    CredentialCreated:
      title: Credential Created
      type: object
      additionalProperties: false
      required: [credential, token]
      properties:
        credential:
          $ref: "#/components/schemas/CredentialView"
        token:
          type: string
          minLength: 1
          description: >-
            Plaintext token shown exactly once. It cannot be recovered from
            the server afterwards.
    CredentialList:
      title: Credential List
      type: object
      additionalProperties: false
      required: [credentials]
      properties:
        credentials:
          type: array
          items:
            $ref: "#/components/schemas/CredentialView"
    Event:
      $ref: "./schemas/event.schema.json"
    EventBatchRequest:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/Event"
    EventBatchResponse:
      type: object
      additionalProperties: false
      required: [results]
      properties:
        results:
          type: array
          minItems: 1
          description: >-
            Exactly one result for each submitted event, in the same order as
            the request array.
          items:
            $ref: "#/components/schemas/EventAcknowledgement"
    EventAcknowledgement:
      type: object
      additionalProperties: false
      required: [event_id, revision, status]
      properties:
        event_id:
          type: string
          format: uuid
        revision:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [accepted, duplicate, stale_revision, rejected]
        error:
          oneOf:
            - $ref: "#/components/schemas/Error"
            - type: "null"
      allOf:
        - if:
            properties:
              status:
                const: rejected
            required: [status]
          then:
            required: [error]
            properties:
              error:
                $ref: "#/components/schemas/Error"
    EventPage:
      type: object
      additionalProperties: false
      required: [data, page, context]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        page:
          $ref: "#/components/schemas/PageMetadata"
        context:
          $ref: "#/components/schemas/QueryContext"
    PageMetadata:
      title: Page Metadata
      type: object
      additionalProperties: false
      required: [page_size, next_cursor]
      properties:
        page_size:
          type: integer
          minimum: 0
          maximum: 200
        next_cursor:
          type: [string, "null"]
          description: Opaque cursor, or null when no further page exists.
    QueryContext:
      title: Query Context
      type: object
      additionalProperties: false
      required: [from, to, timezone, provenance, completeness]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        timezone:
          type: string
        provenance:
          type: array
          items:
            type: string
        completeness:
          type: string
          enum: [complete, partial, unknown]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/Error"
        request_id:
          type: string
          minLength: 1
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          pattern: "^[a-z][a-z0-9_]*$"
        message:
          type: string
          minLength: 1
        details:
          type: object
          additionalProperties: true